Added slowing down time mechanic and more room triggers
Added slowing down time mechanic and more room triggers

+++ 2D Puzzle Game/Assets/Scripts/TimeBar.cs
@@ -9,7 +9,7 @@ public class TimeBar : MonoBehaviour
     public Slider slider;
     public Gradient gradient;
     public Image fill;
-    public void setTime(int time){
+    public void setTime(float time){
         slider.value = time;
 
         fill.color = gradient.Evaluate(slider.normalizedValue);
